+++ src/components/project/edit/TagEdit.tsx
@@ -1,5 +1,6 @@
 import { useState } from 'react';
 import styled from 'styled-components';
+import TagModal from './TagModal';
 interface Props {
     techStack: string[];
     setTechStack: (args: string[]) => void;
@@ -7,13 +8,9 @@ interface Props {
 }
 const LinkEdit = ({ techStack, setTechStack, editMode }: Props) => {
     const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
-    const [modalInput, setModalInput] = useState<string>('');
+
     const handleOpenModal = () => {
         setIsModalOpen(true);
-    };
-
-    const handleCloseModal = () => {
-        if (isModalOpen === true) return setIsModalOpen(false);
     };
 
     return (
@@ -25,7 +22,7 @@ const LinkEdit = ({ techStack, setTechStack, editMode }: Props) => {
                         <Tag>
                             {stack}
                             {editMode && (
-                                <img
+                                <button
                                     onClick={() => {
                                         const newTechStack = [];
                                         for (let i: number = 0; i < techStack.length; i++) {
@@ -35,37 +32,20 @@ const LinkEdit = ({ techStack, setTechStack, editMode }: Props) => {
                                         }
                                         setTechStack(newTechStack);
                                     }}
-                                    src={`${process.env.PUBLIC_URL}/img/close.svg`}
-                                    alt="닫기"
-                                />
+                                >
+                                    <img src={`${process.env.PUBLIC_URL}/img/close.svg`} alt="닫기" />
+                                </button>
                             )}
                         </Tag>
                     ))}
                 {editMode && <Tag onClick={handleOpenModal}>+</Tag>}
-                {isModalOpen && (
-                    <div>
-                        <input
-                            type="text"
-                            value={modalInput}
-                            onChange={(e) => setModalInput(e.target.value)}
-                            placeholder="기술스택을 입력해주세요"
-                        />
-                        <button
-                            onClick={() => {
-                                if (modalInput === '') return;
-                                const newTechStack = [...techStack];
-                                newTechStack.push(modalInput);
-                                setTechStack(newTechStack);
-                                setModalInput('');
-                                handleCloseModal();
-                            }}
-                        >
-                            입력
-                        </button>
-                        <button onClick={handleCloseModal}>모달창 닫기</button>
-                    </div>
-                )}
             </TagDiv>
+            <TagModal
+                isModalOpen={isModalOpen}
+                setIsModalOpen={setIsModalOpen}
+                techStack={techStack}
+                setTechStack={setTechStack}
+            />
         </StackDiv>
     );
 };
@@ -85,7 +65,7 @@ const Tag = styled.div`
     position: relative;
     text-align: center;
     margin: 1%;
-    padding: 1%;
+    padding: 3%;
     height: 3em;
     display: flex;
     justify-content: center;
@@ -94,9 +74,11 @@ const Tag = styled.div`
     color: ${(props) => props.theme.color.sub};
 
     border-radius: 5px;
-    img {
+    button {
         position: absolute;
-        right: 0;
-        top: 0;
+        border: 0px;
+        right: 0px;
+        top: 0px;
+        border: none;
     }
 `;
